+++ Makefile.am
@@ -504,7 +504,6 @@ libnm_core_lib_c_real = \
 	libnm-core/nm-errors.c \
 	libnm-core/nm-keyfile.c \
 	libnm-core/nm-keyfile-utils.c \
-	libnm-core/nm-keyfile-writer.c \
 	libnm-core/nm-property-compare.c \
 	libnm-core/nm-setting.c \
 	libnm-core/nm-simple-connection.c \
